rename to index.d.ts to make it identical to the javascript version
