+++ Dialog/index.d.ts
@@ -1,6 +1,6 @@
 import MaterialComponent from '../MaterialComponent';
 import { VNode } from 'preact';
-import Button from '../Button/Button';
+import Button from '../Button';
 import { MDCFoundation, MDCComponent } from '../MaterialComponentsWeb';
 
 declare interface IDialogProps {
