+++ Slider/index.d.ts
